+++ apps/desktop/src/renderer/routes/_authenticated/components/DashboardNewWorkspaceModal/components/DashboardNewWorkspaceForm/PromptGroup/components/AttachmentButtons/index.ts
@@ -1,0 +1,1 @@
+export { AttachmentButtons } from "./AttachmentButtons";
