+++ apps/desktop/src/renderer/routes/_authenticated/components/DashboardNewWorkspaceModal/components/DashboardNewWorkspaceForm/PromptGroup/components/CompareBaseBranchPicker/index.ts
@@ -1,0 +1,1 @@
+export { CompareBaseBranchPicker } from "./CompareBaseBranchPicker";
