+++ apps/desktop/src/renderer/routes/_authenticated/components/DashboardNewWorkspaceModal/components/DashboardNewWorkspaceForm/PromptGroup/components/ProjectPickerPill/index.ts
@@ -1,0 +1,1 @@
+export { ProjectPickerPill } from "./ProjectPickerPill";
