+++ apps/desktop/src/renderer/routes/_authenticated/components/DashboardNewWorkspaceModal/components/DashboardNewWorkspaceForm/PromptGroup/hooks/useHandleCreate/index.ts
@@ -1,0 +1,1 @@
+export { useHandleCreate } from "./useHandleCreate";
